refactor: Co-locate unit test files with their source code and minor mock fix
This commit makes several organizational improvements and fixes:

1. Co-locate unit tests with their source files:
   - Move component tests to components/ui/
   - Move utility tests to lib/utils/
   - Move config test to lib/config.test.js
   - Move mock client test to lib/mock-minidsp.test.js

2. Rename and reorganize E2E tests:
   - Move tests/e2e/ to integration/e2e/
   - Update playwright.config.js to point to new location
   - Remove redundant meters.spec.js

3. Fix mock client implementation:
   - Add setInputMute and setOutputMute methods to MockMiniDSPClient
   - Ensure proper mute state handling through updateConfig

4. Update imports:
   - Change @/lib/utils to @/lib/tailwind-utils in UI components
   - Update Jest config to find co-located test files

These changes improve code organization by following Jest conventions for unit tests
while keeping integration tests separate. The mock client fix resolves issues with
mute functionality in tests.
