+++ integration/e2e/channels.spec.js
