+++ integration/e2e/meter-rendering.spec.js
