+++ integration/e2e/mute-invert.spec.js
