+++ lib/minidsp-api.js
@@ -174,6 +174,14 @@ export class MockMiniDSPClient extends MiniDSPClient {
   async updateConfig(config) {
     return mockMinidsp.updateConfig(config);
   }
+
+  async setInputMute(index, mute) {
+    return this.updateConfig({ inputs: [{ index, mute }] });
+  }
+
+  async setOutputMute(index, mute) {
+    return this.updateConfig({ outputs: [{ index, mute }] });
+  }
 }
 
 /**
