+++ lib/utils/formatting.test.js
