+++ playwright.config.js
@@ -10,7 +10,7 @@ import { defineConfig } from '@playwright/test';
  * - CI-specific settings for reliability
  */
 export default defineConfig({
-  testDir: './tests/e2e',
+  testDir: './integration/e2e',
   fullyParallel: true,
   forbidOnly: !!process.env.CI,
   retries: process.env.CI ? 2 : 0,
